+++ ErsatzTV.Infrastructure/Health/Checks/FFmpegVersionHealthCheck.cs
@@ -11,15 +11,15 @@ public class FFmpegVersionHealthCheck(
     IHardwareCapabilitiesFactory hardwareCapabilitiesFactory)
     : BaseHealthCheck, IFFmpegVersionHealthCheck
 {
-    private const string BundledVersion = "7.1.1";
-    private const string BundledVersionVaapi = "7.1.1";
-    private const string WindowsVersionPrefix = "n7.1.1";
+    private const string BundledVersion = "8.1.2";
+    private const string BundledVersionVaapi = "8.1.2";
+    private const string WindowsVersionPrefix = "n8.1.2";
 
     public override string Title => "FFmpeg Version";
 
     public async Task<HealthCheckResult> Check(CancellationToken cancellationToken)
     {
-        var link = new HealthCheckLink("https://github.com/ErsatzTV/ErsatzTV-ffmpeg/releases/tag/7.1.1");
+        var link = new HealthCheckLink("https://github.com/ErsatzTV/ErsatzTV-ffmpeg/releases/tag/8.1.2");
 
         Option<ConfigElement> maybeFFmpegPath =
             await configElementRepository.GetConfigElement(ConfigElementKey.FFmpegPath, cancellationToken);
@@ -82,21 +82,22 @@ public class FFmpegVersionHealthCheck(
         if (version.StartsWith("3.", StringComparison.OrdinalIgnoreCase) ||
             version.StartsWith("4.", StringComparison.OrdinalIgnoreCase) ||
             version.StartsWith("5.", StringComparison.OrdinalIgnoreCase) ||
-            version.StartsWith("6.", StringComparison.OrdinalIgnoreCase))
+            version.StartsWith("6.", StringComparison.OrdinalIgnoreCase) ||
+            version.StartsWith("7.", StringComparison.OrdinalIgnoreCase))
         {
             return FailResult(
-                $"{app} version {version} is too old; please install 7.1.1!",
+                $"{app} version {version} is too old; please install 8.1.2!",
                 $"{app} version is too old",
                 link);
         }
 
-        if (!version.StartsWith("7.1.1", StringComparison.OrdinalIgnoreCase) &&
+        if (!version.StartsWith("8.1.2", StringComparison.OrdinalIgnoreCase) &&
             !version.StartsWith(WindowsVersionPrefix, StringComparison.OrdinalIgnoreCase) &&
             version != BundledVersion &&
             version != BundledVersionVaapi)
         {
             return WarningResult(
-                $"{app} version {version} is unexpected and may have problems; please install 7.1.1!",
+                $"{app} version {version} is unexpected and may have problems; please install 8.1.2!",
                 $"{app} version is unexpected",
                 link);
         }